[agent] fix(cli): isolate restore audit persistence
